Add files via upload
The Sensor-Based Traffic Light System integrates an ultrasonic sensor (HC-SR04) with a traditional traffic light to improve road safety and efficiency. The system follows a normal traffic light cycle with green, yellow, and red signals, but it also incorporates a button and ultrasonic sensor to detect emergency situations or obstacles. If no vehicle o is detected within a 100 cm range, or if the emergency button is pressed, the system immediately activates the red light to stop traffic, ensuring safety. Once the  button is released, the system automatically resumes the normal traffic cycle. This smart approach enhances real-time traffic management, reducing accidents and improving road efficiency